requirements.txt
streamlit
google.cloud
google-cloud-storage
